Add files via upload
This is the latest update since 11/20/16. It is still a work in progress, but most of the memory registers and text from the book have been added. It still needs some formatting since I finally started copying text from a website I found online (it was getting tiring typing in everything). Anyway enjoy!